[Routing] Refactoring
Almost identical Generate()
from CarDirectionsEngine and PedestrianDirectionsEngine
moved to parent class.

MakeTurnAnnotation() - became method of DirectionsEngine.

All VehicleType dependent logic was incapsulated in virtual methods:
- GetTurnDirection()
- FixupTurns().

According existing functions became methods of
CarDirectionsEngine and PedestrianDirectionsEngine.

VehicleType dependent functions moved
from turns_generator to according DirectionsEngines.
