+++ routing/directions_engine.cpp
@@ -1,6 +1,7 @@
 #include "routing/directions_engine.hpp"
 #include "routing/fake_feature_ids.hpp"
 #include "routing/routing_helpers.hpp"
+#include "routing/routing_callbacks.hpp"
 
 #include "traffic/traffic_info.hpp"
 
@@ -229,4 +230,182 @@ void DirectionsEngine::FillPathSegmentsAndAdjacentEdgesMap(
     startSegId = kInvalidSegId;
   }
 }
+
+bool DirectionsEngine::Generate(IndexRoadGraph const & graph,
+                                vector<geometry::PointWithAltitude> const & path,
+                                base::Cancellable const & cancellable, Route::TTurns & turns,
+                                Route::TStreets & streetNames,
+                                vector<geometry::PointWithAltitude> & routeGeometry,
+                                vector<Segment> & segments)
+{
+  CHECK(m_numMwmIds, ());
+
+  m_adjacentEdges.clear();
+  m_pathSegments.clear();
+  turns.clear();
+  streetNames.clear();
+  segments.clear();
+
+  IndexRoadGraph::EdgeVector routeEdges;
+
+  CHECK_NOT_EQUAL(m_vehicleType, VehicleType::Count, (m_vehicleType));
+
+  if (m_vehicleType == VehicleType::Transit)
+  {
+    routeGeometry = path;
+    graph.GetRouteSegments(segments);
+    graph.GetRouteEdges(routeEdges);
+    turns.emplace_back(routeEdges.size(), turns::PedestrianDirection::ReachedYourDestination);
+    return true;
+  }
+
+  routeGeometry.clear();
+
+  if (path.size() <= 1)
+    return false;
+
+  graph.GetRouteEdges(routeEdges);
+
+  if (routeEdges.empty())
+    return false;
+
+  if (cancellable.IsCancelled())
+    return false;
+
+  FillPathSegmentsAndAdjacentEdgesMap(graph, path, routeEdges, cancellable);
+
+  if (cancellable.IsCancelled())
+    return false;
+
+  auto const res = MakeTurnAnnotation(routeEdges, cancellable,
+                                      routeGeometry, turns, streetNames, segments);
+
+  if (res != RouterResultCode::NoError)
+    return false;
+
+  // In case of bicycle routing |m_pathSegments| may have an empty
+  // |LoadedPathSegment::m_segments| fields. In that case |segments| is empty
+  // so size of |segments| is not equal to size of |routeEdges|.
+  if (!segments.empty())
+    CHECK_EQUAL(segments.size(), routeEdges.size(), ());
+  return true;
+}
+
+/*!
+ * \brief Compute turn and time estimation structs for the abstract route result.
+ * \param result abstract routing result to annotate.
+ * \param delegate Routing callbacks delegate.
+ * \param points Storage for unpacked points of the path.
+ * \param turnsDir output turns annotation storage.
+ * \param streets output street names along the path.
+ * \param segments route segments.
+ * \return routing operation result code.
+ */
+
+// Normally loadedSegments structure is:
+// - Start point. Fake loop LoadedPathSegment with 1 segment of zero length.
+// - Straight jump from start point to the beginning of real route. LoadedPathSegment with 1 segment.
+// - Real route. N LoadedPathSegments, each with arbitrary amount of segments. N >= 1.
+// - Straight jump from the end of real route to finish point. LoadedPathSegment with 1 segment.
+// - Finish point. Fake loop LoadedPathSegment with 1 segment of zero length.
+// So minimal amount of segments is 5.
+
+// Resulting structure of turnsDir:
+// No Turn for 0th segment (no ingoing). m_index == 0.
+// No Turn for 1st segment (ingoing fake loop) - at start point. m_index == 1.
+// No Turn for 2nd (ingoing == jump) - at beginning of real route. m_index == 2.
+// Possible turn for next N-1 segments. m_index >= 3.
+// No Turn for (2 + N + 1)th segment (outgoing jump) - at finish point. m_index = 3 + M.
+// No Turn for (2 + N + 2)th segment (outgoing fake loop) - at finish point. m_index == 4 + M.
+// Added ReachedYourDestination - at finish point. m_index == 4 + M.
+// Where M - total amount of all segments from all LoadedPathSegments (size of |segments|).
+// So minimum m_index of ReachedYourDestination is 5 (real route with single segment),
+// and minimal |turnsDir| is - single ReachedYourDestination with m_index == 5.
+
+RouterResultCode DirectionsEngine::MakeTurnAnnotation(IndexRoadGraph::EdgeVector const & routeEdges,
+                                                      base::Cancellable const & cancellable,
+                                                      std::vector<geometry::PointWithAltitude> & junctions,
+                                                      Route::TTurns & turnsDir, Route::TStreets & streets,
+                                                      std::vector<Segment> & segments)
+{
+  RoutingEngineResult result(routeEdges, m_adjacentEdges, m_pathSegments);
+
+  LOG(LDEBUG, ("Shortest path length:", result.GetPathLength()));
+
+  if (cancellable.IsCancelled())
+    return RouterResultCode::Cancelled;
+
+  size_t skipTurnSegments = 0;
+  auto const & loadedSegments = result.GetSegments();
+  segments.reserve(loadedSegments.size());
+
+  RoutingSettings const vehicleSettings = GetRoutingSettings(m_vehicleType);
+
+  for (auto loadedSegmentIt = loadedSegments.cbegin(); loadedSegmentIt != loadedSegments.cend(); ++loadedSegmentIt)
+  {
+    CHECK(loadedSegmentIt->IsValid(), ());
+
+    // Street names contain empty names too for avoiding of freezing of old street name while
+    // moving along unnamed street.
+    streets.emplace_back(max(junctions.size(), static_cast<size_t>(1)) - 1, loadedSegmentIt->m_name);
+
+    // Turns information.
+    if (!junctions.empty() && skipTurnSegments == 0)
+    {
+      auto const outgoingSegmentDist = distance(loadedSegments.begin(), loadedSegmentIt);
+      CHECK_GREATER(outgoingSegmentDist, 0, ());
+      auto const outgoingSegmentIndex = static_cast<size_t>(outgoingSegmentDist);
+
+      TurnItem turnItem;
+      turnItem.m_index = static_cast<uint32_t>(junctions.size() - 1);
+
+      skipTurnSegments = GetTurnDirection(result, outgoingSegmentIndex, *m_numMwmIds, vehicleSettings, turnItem);
+
+      if (!turnItem.IsTurnNone())
+        turnsDir.push_back(move(turnItem));
+    }
+
+    if (skipTurnSegments > 0)
+      --skipTurnSegments;
+
+    // Path geometry.
+    CHECK_GREATER_OR_EQUAL(loadedSegmentIt->m_path.size(), 2, ());
+    // Note. Every LoadedPathSegment in TUnpackedPathSegments contains LoadedPathSegment::m_path
+    // of several Junctions. Last PointWithAltitude in a LoadedPathSegment::m_path is equal to first
+    // junction in next LoadedPathSegment::m_path in vector TUnpackedPathSegments:
+    // *---*---*---*---*       *---*           *---*---*---*
+    //                 *---*---*   *---*---*---*
+    // To prevent having repetitions in |junctions| list it's necessary to take the first point only
+    // from the first item of |loadedSegments|. The beginning should be ignored for the rest
+    // |m_path|.
+    junctions.insert(junctions.end(), loadedSegmentIt == loadedSegments.cbegin()
+                                          ? loadedSegmentIt->m_path.cbegin()
+                                          : loadedSegmentIt->m_path.cbegin() + 1,
+                     loadedSegmentIt->m_path.cend());
+    segments.insert(segments.end(), loadedSegmentIt->m_segments.cbegin(),
+                    loadedSegmentIt->m_segments.cend());
+  }
+
+  // Path found. Points will be replaced by start and end edges junctions.
+  if (junctions.size() == 1)
+    junctions.push_back(junctions.front());
+
+  if (junctions.size() < 2)
+    return RouterResultCode::RouteNotFound;
+
+  junctions.front() = result.GetStartPoint();
+  junctions.back() = result.GetEndPoint();
+
+  FixupTurns(junctions, turnsDir);
+
+/*
+  CHECK_EQUAL(
+      routeGeometry.size(), pathSize,
+      ("routeGeometry and path have different sizes. routeGeometry size:", routeGeometry.size(),
+       "path size:", pathSize, "segments size:", segments.size(), "routeEdges size:",
+       routeEdges.size(), "resultGraph.GetSegments() size:", resultGraph.GetSegments().size()));
+*/
+  return RouterResultCode::NoError;
+}
+
 }  // namespace routing
