+++ routing/directions_engine.hpp
@@ -19,6 +19,15 @@
 
 namespace routing
 {
+
+namespace turns
+{
+class IRoutingResult;
+struct TurnItem;
+}
+
+enum class RouterResultCode;
+
 class DirectionsEngine
 {
 public:
@@ -35,17 +44,28 @@ public:
   /// \brief Generates all args which are passed by reference.
   /// \param path is points of the route. It should not be empty.
   /// \returns true if fields passed by reference are filled correctly and false otherwise.
-  virtual bool Generate(IndexRoadGraph const & graph,
-                        std::vector<geometry::PointWithAltitude> const & path,
-                        base::Cancellable const & cancellable, Route::TTurns & turns,
-                        Route::TStreets & streetNames,
-                        std::vector<geometry::PointWithAltitude> & routeGeometry,
-                        std::vector<Segment> & segments) = 0;
+  bool Generate(IndexRoadGraph const & graph,
+                std::vector<geometry::PointWithAltitude> const & path,
+                base::Cancellable const & cancellable, Route::TTurns & turns,
+                Route::TStreets & streetNames,
+                std::vector<geometry::PointWithAltitude> & routeGeometry,
+                std::vector<Segment> & segments);
   void Clear();
 
   void SetVehicleType(VehicleType const & vehicleType) { m_vehicleType = vehicleType; }
 
 protected:
+  /*!
+  * \brief GetTurnDirection makes a primary decision about turns on the route.
+  * \param outgoingSegmentIndex index of an outgoing segments in vector result.GetSegments().
+  * \param turn is used for keeping the result of turn calculation.
+  */
+  virtual size_t GetTurnDirection(turns::IRoutingResult const & result, size_t const outgoingSegmentIndex,
+                                  NumMwmIds const & numMwmIds,
+                                  RoutingSettings const & vehicleSettings, turns::TurnItem & turn) = 0;
+  virtual void FixupTurns(std::vector<geometry::PointWithAltitude> const & junctions,
+                          Route::TTurns & turnsDir) = 0;
+
   FeaturesLoaderGuard & GetLoader(MwmSet::MwmId const & id);
   void LoadPathAttributes(FeatureID const & featureId, LoadedPathSegment & pathSegment);
   void GetSegmentRangeAndAdjacentEdges(IRoadGraph::EdgeListT const & outgoingEdges,
@@ -70,5 +90,12 @@ protected:
   std::shared_ptr<NumMwmIds> m_numMwmIds;
   std::unique_ptr<FeaturesLoaderGuard> m_loader;
   VehicleType m_vehicleType = VehicleType::Count;
+
+private:
+  RouterResultCode MakeTurnAnnotation(IndexRoadGraph::EdgeVector const & routeEdges,
+                                      base::Cancellable const & cancellable,
+                                      std::vector<geometry::PointWithAltitude> & junctions,
+                                      Route::TTurns & turnsDir, Route::TStreets & streets,
+                                      std::vector<Segment> & segments);
 };
 }  // namespace routing
